Re-organized directories, got factory working
