--- Picker/picker.py
@@ -1,11 +1,0 @@
-
-
-class Picker(object):
-"""Defines the Picker class. Each picker has a different method for choosing the numbers in a
-drawing. Many pickers will use past drawings to """
-
-    def __init__(self):
-        self.picks = []
-
-    def Pick(pastDrawings, numPicks, altWhites=True):
-        return self.picks
