--- __init__.py
@@ -1,2 +1,0 @@
-MAX_WHITE = 69
-MAX_RED = 26
