+++ src/picker/__init__.py
@@ -1,0 +1,9 @@
+
+
+class Picker(object):
+
+    def __init__(self):
+        self.picks = []
+
+    def Pick(pastDrawings, numPicks, altReds=True):
+        return self.picks
