+++ src/picker/leastRecentPicker.py
@@ -1,0 +1,15 @@
+
+#from src.utilities.metrics import leastRecentReds, leastRecentWhites
+from picker import Picker
+
+class LeastRecentPicker(Picker):
+
+    def __init__(self):
+        super(RandomPicker, self).__init__()
+
+    def Pick(pastDrawings, numPicks, altReds=True):
+    	return None
+        #lrReds = leastRecentReds(pastDrawings)
+        #lrWhites = leastRecentWhites(pastDrawings)
+
+
